productBackend/requirements.txt
Django
mysqlclient
python-dotenv
django-cors-headers
html5lib
numpy
Pillow
djangorestframework
requests
dataclasses
tweepy==4.14.0
djangorestframework-simplejwt